Update spamhaus.sh
I have changed this to accept a few flags to make it runnable from cron quietly, runnable from rc.local with cached content, cached spamhous drop.lasso, and have allowed for a dry run of the downloaded file.

The previous version would automatically make the mail server unprotected should there be an error in downloading the spamhaus file.

The downloaded file can be further validated before it is moved but at this time, this is good enough. If the spamhaus file is empty, we ignore it. It should probably generate an email but I have not gotten to that yet.